total weekly hours and calcuation of daily values
